chore: remove JSON error handler from examples
This confuses and complicates the Express example.
